Folder reorganization bug fixes
