+++ Demos/Haltable/HaltableMicDemo.py
@@ -1,5 +1,5 @@
 from threading import Thread
-from Demos.IMicDemo import IMicDemo
+from Demos.Interfaces.IMicDemo import IMicDemo
 
 
 # Purely for documentation purposes
