+++ Demos/Haltable/HaltableVoiceDemo.py
@@ -1,5 +1,5 @@
 import speech_recognition as sr
-from Demos.IRobotDemo import IRobotDemo
+from Demos.Interfaces.IRobotDemo import IRobotDemo
 
 
 class HaltableVoiceDemo(IRobotDemo):
@@ -20,7 +20,6 @@ class HaltableVoiceDemo(IRobotDemo):
                 print("Speak a command.")
                 r.adjust_for_ambient_noise(source, duration=0.2)
                 audio = r.listen(source)
-                # text = r.recognize_google(audio)
 
                 response = {
                     "success": True,
@@ -28,18 +27,13 @@ class HaltableVoiceDemo(IRobotDemo):
                     "transcription": None
                 }
 
-                # try recognizing the speech in the recording
-                # if a RequestError or UnknownValueError exception is caught,
-                #     update the response object accordingly
                 try:
                     response["transcription"] = r.recognize_google(audio)
                 except sr.RequestError:
-                    # API was unreachable or unresponsive
                     response["success"] = False
                     response["error"] = "API unavailable"
                     continue
                 except sr.UnknownValueError:
-                    # speech was unintelligible
                     response["error"] = "Unable to recognize speech"
                     continue
