+++ Handlers/PerformanceHandler.py
@@ -1,6 +1,10 @@
 import time
 import threading
 
+from pythonosc import udp_client
+from pythonosc import dispatcher
+from pythonosc import osc_server
+
 
 class PerformanceHandler:
     def __init__(self, robotHandler, is_lab_work=True):
