+++ Performance.py
@@ -13,7 +13,7 @@ class Performance:
         self.voiceDemo = HaltableVoiceDemo(robotHandler, is_lab_work)
         self.micDemo = HaltableMicDemo(
             robotHandler, is_lab_work, robots_already_awake=True)
-        self.visionTrackerDemo = VisionTrackerDemo(robotHandler, is_lab_work)
+        # self.visionTrackerDemo = VisionTrackerDemo(robotHandler, is_lab_work)
 
         self.mainThread = Thread()
 
@@ -38,11 +38,18 @@ class Performance:
         self.micDemo.kill()
         mainThread.join()
 
-        # 0:31 - END ------------- 2 -
+        # 0:31 - 3:30 ------------ 2 -
         # Xylophone Mimicking (Audio)
         # ----------------------------
-        mainThread = Thread(target=self.visionTrackerDemo.start)
+        # mainThread = Thread(target=self.visionTrackerDemo.start)
+        # mainThread.start()
+
+        # END--------------------- 2 -
+        # Xylophone Mimicking (Audio)
+        # ----------------------------
+        mainThread = Thread(target=self.voiceDemo.start)
         mainThread.start()
+        mainThread.join()
 
 
 robotHandler = RobotHandler(is_lab_work=False)
